+++ src/testing/benchmarks/SingleFFTBench.cu
@@ -37,13 +37,13 @@ int main(){
   int sample_size = 200;
   int warmup_samples = 5;
 
-  std::vector<int> fft_length;
+  std::vector<long long> fft_length;
   std::vector<double> avg_runtime;
   std::vector<double> sigma_runtime;
 
   std::optional<std::string> error_mess;
 
-  int length = 16 * 8;
+  long long length = 16 * 8;
   for(int i=8; i<=log_length_max; i++){
     length = length * 2;
     fft_length.push_back(length);
